AABB/AABB BVH tree with post-filtering.
Removing the triangles. Will index instead the bounding box of the
geometry and apply filters in the scorer. Reduces the size of the index
such that the mmap temporary file is no longer need. We no longer need
the quantized i32 and can instead use f64 which is what is expected for
geo search.

Additionally, this creates a clear path to 2d-on-sphere to address polar
distortion. Constrained triangulation on a sphere is not readily
available in open source, so projection for 2d tessellation and
unprojection would have been required to support spheres.

Sphere predicates are going to be more easily implemented using logic
that is already out there in existing open source libraries, most
notably S2. The BVH tree already supports 3d bounding volumes, but we
will need boundary generator that accounts for great circle arc bulges
for indexing, and a separate set of predicates.

Removed iTriangle.

Also removed attempts at cache-line alignment and making no assumptions
about position in memory.

Retaining the memory based partitioning using select_nth_unstable and
will want to preserve it even if moving to a different partition
algorithm at it simplifies the implementation of partitioning. Records
are still of a fixed size to account for allowing other programs to read
the files and some potential compression strategies.