Drop i32 and commit to f64.
For the planar index, 10,000,000 docs is 360MB in memory at merge time.
The same docs tessellated into an average of 24 triangles per doc with
the i32 record would have been 7GB. i32 made sense then since the
integers delta compressed better than doubles XOR compress and they are
half this size. It no longer makes sense and we do not have trouble the
user with quantization.

f64 is suitable for astronomy. If 2^53 can map the universe, it should
be able to map the things within it. If a user does want integers,
53-bit mantissa preserves them and the BVH tree is only doing
comparisons, not arithmetic, so the BVH logic is unaffected.

Furthermore, we're already storing the Geometry as f64, so we're already
at the 53-bit mantissa limit so to make our BVH tree meaningfully
generic over coordinates we'd have to make our geometry storage generic
as well.